Create DeleteNoteModal to handle Note deletion
Needed to use a Modal to inform user note deletion
is destructive and can't be undone. Also need to
refactor all the Models to add a new isDeleted
property to mark documents that are deleted.
Manual deletion is time consuming. So just mark all
or just the main Note document as deleted and
later, run a job that will purge database of all
Note s3 Files, DB File documents, and main Note
DB document at a later time.