Start refactoring /notes/[noteID]/edit page for noteFiles conditional rendering
The issue is that if the Note has no attached File,
it's pointless to allow the use to have a conversation
with the LLM if there's context available for the LLM
to query. Need to conditionally render sections of
the /edit page based on whether the current Note has
any attached .txt files.